delete confirm functionality and security on add/del/profile

+++ services/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 from .models import Services
 from .forms import ServicesForm
@@ -31,8 +32,13 @@ def service_detail(request, services_id):
     return render(request, 'services/service_detail.html', context)
 
 
+@login_required
 def add_service(request):
     """ Add a service to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners cad do that!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ServicesForm(request.POST, request.FILES)
         if form.is_valid():
@@ -53,8 +59,13 @@ def add_service(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_service(request, services_id):
     """ Edit a service in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners cad do that!')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Services, pk=services_id)
     if request.method == 'POST':
         form = ServicesForm(request.POST, request.FILES, instance=service)
@@ -78,8 +89,13 @@ def edit_service(request, services_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_service(request, services_id):
     """Delete service from store"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners cad do that!')
+        return redirect(reverse('home'))
+
     service = get_object_or_404(Services, pk=services_id)
     service.delete()
     messages.success(request, 'Service deleted!')
